Full Basic Student
This commit has the full functionality of the basic student. You are
able to add students, delete students, inspect a students information by
clicking on their name & edit their information. This commit has a small
amount of additional code which is the start of adding other classes to
the student functionality :+1: