Added Gender to Student Model
This commit allows for genders to be included on the student object.
Currently only the add student function is the only interaction with
gender but the object relation is stored in the student object. ATM it
the gender object doesn't store a link to the students but this may not
be needed.